+++ docs/versionhistory.rst
@@ -5,6 +5,7 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
 **UNRELEASED**
 
+- Added introspection of running tasks via ``anyio.get_running_tasks()``
 - Fixed mishandling of large buffers by ``BaseSocket.sendall()``
 
 **1.0.0b1**